+++ src/server/ResponseHandlers.hpp
@@ -23,15 +23,14 @@ public:
                std::pair<FileDescriptor const *, CgiDelegate *> > &cgis,
       FileDescriptor const *client_fd, char **envp);
 
-private:
-  static Response delete_method(const Target &target, Response response,
+  static Response delete_method_response(const Target &target, Response response,
                                 const ServerConfig *config,
                                 const RouteRule *rule, char **envp);
-  static Response post_method(const Target &target, Response response,
+  static Response post_method_response(const Target &target, Response response,
                               const ClientSession *client,
                               const RouteRule *rule, const Request *request,
                               Session *session, char **envp);
-  static Response get_method(Target target, Response response,
+  static Response get_method_response(Target target, Response response,
                              const ServerConfig *config, const RouteRule *rule,
                              const Request *request, char **envp);
 };
